Make consecutive resend configurable
Added configuration options MnMaxHaConsecutiveResends and
MnMaxCnConsecutiveResends in order to make BU resend configurable.

MnMaxHaConsecutiveResends replaces the MAX_CONSECUTIVE_RESENDS macro.
By default, the previous behavior is enfoced: MnMaxHaConsecutiveResends
is 5 and MnMaxCnConsecutiveResends is 0 (resend indefinitely).

+++ src/bul.h
@@ -40,7 +40,7 @@ struct bulentry {
 	struct timespec delay;		/* call back time in ms*/
 	struct timespec expires;        /* Absolute time for timer expire */
 	struct timespec hard_expire;    /* Absolute bulentry expiry time */
-	int consecutive_resends;	/* Number of consecutive BU's resent */
+	uint16_t consecutive_resends;	/* Number of consecutive BU's resent */
 	int8_t coa_changed;
 	uint8_t wait_ack;      		/* WAIT / READY */
 	uint8_t xfrm_state;
